+++ NebulaCompatibilityAssist/src/Plugin.cs
@@ -21,7 +21,7 @@ namespace NebulaCompatibilityAssist
     {
         public const string GUID = "NebulaCompatibilityAssist";
         public const string NAME = "NebulaCompatibilityAssist";
-        public const string VERSION = "0.4.18";
+        public const string VERSION = "0.4.19";
 
         public static Plugin Instance { get; private set; }
         public Harmony Harmony { get; private set; }